requirements.txt
asgiref==3.5.0
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.1
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==3.2.11
django-cors-headers==3.11.0
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.1.0
pyotp==2.6.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.0.1
urllib3==1.26.12
xlrd==2.0.1
zope.interface==5.4.0
